Recipes_Ingredients/recipe-chat-system/backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pgvector==0.2.4

# Model & embedding stack
sentence-transformers==5.0.0
huggingface-hub==0.33.2

ollama==0.5.1

# Validation & settings
pydantic==2.11.7
pydantic-settings==2.7.1
email-validator==2.2.0
python-multipart==0.0.20
bcrypt==4.2.1
python-dotenv==1.1.1